requirements.txt
pytest==7.4.3
playwright==1.40.0
allure-pytest==2.13.2
python-dotenv==1.0.0
pytest-html==4.1.1
pytest-xdist==3.3.1
